requirements.txt
Django==2.2.5
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.4.0
Markdown==3.3.3
numpy==1.19.5
pandas==1.1.5
psycopg2==2.7.6.1
psycopg2-binary==2.8
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
